+++ src/utils/config_validater.py
@@ -2,6 +2,7 @@ from omegaconf import DictConfig
 from pathlib import Path
 from typing import Tuple, Dict, List
 from utils import hydra_utils
+from src.functions import solver_interfaces
 
 from hydra.core.hydra_config import HydraConfig
 import os
@@ -31,7 +32,7 @@ def validate_config(cfg: DictConfig):
     if cfg.config_validation.validate_solver:
         solver_res, solver_valid = check_solver_config(cfg.solver)
         all_validation_results.append(solver_valid)
-        print_solver_report(solver_res, solver_valid)
+        print_solver_report(solver_res, solver_valid,cfg)
     else:
         print(f"{YELLOW}WARNING: Validation of solver config disabeled!{RESET}")
 
@@ -61,6 +62,7 @@ def experiment_root_exists(cfg: DictConfig) -> bool:
     if os.path.exists(root_dir):
         root_dir = hydra_utils.get_unique_dir_name(root_dir)
 
+
     print(root_dir)
 
 
@@ -68,11 +70,54 @@ def check_solver_config(config) -> Tuple[dict, bool]:
     # Things for solver to check:
     # - config.path exists
     # - unique solver names
-    return None, True
+
+    solver_config_validation_results = {
+        "solver_path_exists": False,
+        "interface_supported": False,
+    }
 
 
-def print_solver_report(solver_validation_results: dict, is_valid: bool):
-    print(f"{RED} SOLVER-REPORT{RESET}")
+
+    # check if solver path exists, if not return false
+    if Path(config.path).exists():
+        solver_config_validation_results["solver_path_exists"] = True
+
+     # Check if the solver interface is supported
+    if config['interface'] in solver_interfaces.interfaces_dict.keys():
+        solver_config_validation_results["interface_supported"] = True
+
+
+    # Run solver on dummy instances to check interface and if solver is working
+
+
+
+    # Determine if all checks pass
+    valid = all(solver_config_validation_results.values())
+    return solver_config_validation_results, valid
+
+def print_solver_report(solver_validation_results: dict, is_valid: bool, cfg: DictConfig):
+    """
+    Prints a detailed benchmark validation report based on the validation results.
+
+    Args:
+        benchmark_validation_results (dict): The results dictionary from the benchmark validation.
+        is_valid (bool): Indicates if the benchmark configuration is valid.
+    """
+    if is_valid:
+        print(f"{GREEN}✔ Solver configuration is valid.{RESET}")
+    else:
+        print(f"{RED}✘ Solver configuration is not valid.{RESET}")
+
+    # Detailed report of each check
+    if solver_validation_results["solver_path_exists"]:
+        print(f"    {GREEN}✔ Solver executable path exists{RESET}")
+    else:
+        print(f"    {RED}✘ Solver executable path exists{RESET} ({cfg.solver.path})")
+
+    if solver_validation_results["interface_supported"]:
+        print(f"    {GREEN}✔ Interface {cfg.solver.interface} not supported{RESET}")
+    else:
+        print(f"    {RED}✘ Interface {cfg.solver.interface} supported{RESET}")
 
 
 def check_benchmark_config(config: Dict) -> Tuple[Dict[str, bool], bool]: